@@ -40,10 +40,11 @@
     "prepublish": "yarn run build"
   },
   "dependencies": {
-    "babel-runtime": "^6.26.0",
     "@babel/runtime-corejs2": "^7.7.7",
+    "@vue/compat": "^3.0.0",
+    "babel-runtime": "^6.26.0",
     "froala-editor": "^4.0.17",
-    "vue": "2.6.6"
+    "vue": "^3.2.0"
   },
   "devDependencies": {
     "@babel/cli": "^7.2.3",
@@ -54,6 +55,7 @@
     "@babel/preset-env": "^7.3.4",
     "@babel/runtime": "^7.3.4",
     "@babel/runtime-corejs2": "^7.3.4",
+    "@vue/compiler-sfc": "3.2",
     "babel-helpers": "^6.24.1",
     "babel-loader": "^8.0.5",
     "connect-history-api-fallback": "^1.1.0",
@@ -62,7 +64,6 @@
     "css-loader": "^2.1.0",
     "eventsource-polyfill": "^0.9.6",
     "express": "^4.15.2",
-    "mini-css-extract-plugin": "1.6.2",
     "file-loader": "^3.0.1",
     "function-bind": "^1.0.2",
     "html-webpack-plugin": "^3.2.0",
@@ -71,6 +72,7 @@
     "isparta-loader": "^2.0.0",
     "jasmine-core": "^3.3.0",
     "lolex": "^3.1.0",
+    "mini-css-extract-plugin": "1.6.2",
     "nightwatch": "^1.0.19",
     "ora": "^3.1.0",
     "shelljs": "^0.8.3",
@@ -78,9 +80,9 @@
     "url-loader": "^1.1.2",
     "vue-hot-reload-api": "^2.0.11",
     "vue-html-loader": "^1.2.4",
-    "vue-loader": "15.6.2",
+    "vue-loader": "^16.8.3",
     "vue-style-loader": "^4.1.2",
-    "vue-template-compiler": "2.6.6",
+    "vue-template-compiler": "^2.7.14",
     "webpack": "^4.29.3",
     "webpack-cli": "3.2.3",
     "webpack-dev-middleware": "^3.5.2",
@@ -98,7 +100,7 @@
   },
   "greenkeeper": {
     "ignore": [
-      "extract-text-webpack-plugin",
+      "mini-css-extract-plugin",
       "karma-webpack",
       "webpack",
       "webpack-merge"
